+++ network-api/networkapi/wagtailpages/tests/research_hub/test_library_page.py
@@ -378,78 +378,6 @@ class TestResearchLibraryPage(research_test_base.ResearchHubTestCase):
         newest_first_detail_pages = list(newest_first_response.context["research_detail_pages"])
         self.assertEqual(default_sort_detail_pages, newest_first_detail_pages)
 
-    def test_research_author_profile_in_options(self):
-        detail_page = detail_page_factory.ResearchDetailPageFactory(
-            parent=self.library_page,
-        )
-
-        response = self.client.get(self.library_page.url)
-
-        author_option_values = [i["value"] for i in response.context["author_options"]]
-        self.assertIn(
-            detail_page.research_authors.first().author_profile.id,
-            author_option_values,
-        )
-
-    def test_non_research_author_profile_not_in_options(self):
-        profile = profiles_factory.ProfileFactory()
-
-        response = self.client.get(self.library_page.url)
-
-        author_option_values = [i["value"] for i in response.context["author_options"]]
-        self.assertNotIn(
-            profile.id,
-            author_option_values,
-        )
-
-    def test_research_author_in_context_aliased_detail_page_fr(self):
-        """
-        After the treesync, there are alias pages in the non-default locales. But,
-        before the pages are translated (a manual action) the related models like author
-        are still the ones from the default locale.
-        """
-        detail_page_en = detail_page_factory.ResearchDetailPageFactory(
-            parent=self.library_page,
-        )
-        profile_en = detail_page_en.research_authors.first().author_profile
-        self.synchronize_tree()
-        translation.activate(self.fr_locale.language_code)
-
-        response = self.client.get(self.library_page.localized.url)
-
-        author_option_values = [i["value"] for i in response.context["author_options"]]
-        self.assertIn(
-            profile_en.id,
-            author_option_values,
-        )
-
-    def test_research_author_in_context_translated_detail_page_fr(self):
-        """
-        When a profile for the active locale exists, pass that one to the context.
-
-        Profiles are not necessarily people, so they might have translated names.
-        """
-        detail_page_en = detail_page_factory.ResearchDetailPageFactory(
-            parent=self.library_page,
-        )
-        profile_en = detail_page_en.research_authors.first().author_profile
-        self.synchronize_tree()
-        detail_page_fr = research_test_utils.translate_detail_page(detail_page_en, self.fr_locale)
-        profile_fr = detail_page_fr.research_authors.first().author_profile
-        translation.activate(self.fr_locale.language_code)
-
-        response = self.client.get(self.library_page.localized.url)
-
-        author_option_values = [i["value"] for i in response.context["author_options"]]
-        self.assertNotIn(
-            profile_en.id,
-            author_option_values,
-        )
-        self.assertIn(
-            profile_fr.id,
-            author_option_values,
-        )
-
     def test_filter_author_profile(self):
         detail_page_1 = detail_page_factory.ResearchDetailPageFactory(
             parent=self.library_page,
@@ -537,57 +465,6 @@ class TestResearchLibraryPage(research_test_base.ResearchHubTestCase):
         self.assertNotIn(detail_page_1, research_detail_pages)
         self.assertNotIn(detail_page_2, research_detail_pages)
 
-    def test_research_topics_in_options(self):
-        topic_1 = taxonomies_factory.ResearchTopicFactory()
-        topic_2 = taxonomies_factory.ResearchTopicFactory()
-
-        response = self.client.get(self.library_page.url)
-
-        topic_option_values = [i["value"] for i in response.context["topic_options"]]
-        self.assertEqual(len(topic_option_values), 2)
-        self.assertIn(topic_1.id, topic_option_values)
-        self.assertIn(topic_2.id, topic_option_values)
-
-    def test_topic_in_options_matches_active_locale(self):
-        topic_en = taxonomies_factory.ResearchTopicFactory()
-        topic_fr = topic_en.copy_for_translation(self.fr_locale)
-        topic_fr.save()
-
-        response_en = self.client.get(self.library_page.localized.url)
-        translation.activate(self.fr_locale.language_code)
-        response_fr = self.client.get(self.library_page.localized.url)
-
-        topic_option_values_en = [i["value"] for i in response_en.context["topic_options"]]
-        self.assertEqual(len(topic_option_values_en), 1)
-        self.assertIn(topic_en.id, topic_option_values_en)
-        self.assertNotIn(topic_fr.id, topic_option_values_en)
-        topic_option_values_fr = [i["value"] for i in response_fr.context["topic_options"]]
-        self.assertEqual(len(topic_option_values_fr), 1)
-        self.assertNotIn(topic_en.id, topic_option_values_fr)
-        self.assertIn(topic_fr.id, topic_option_values_fr)
-
-    def test_localized_topic_options(self):
-        """
-        Use active locales version of topic if available.
-
-        If no translation is available for a given topic, display the default locale
-        topic.
-
-        """
-        topic_1_en = taxonomies_factory.ResearchTopicFactory()
-        topic_1_fr = topic_1_en.copy_for_translation(self.fr_locale)
-        topic_1_fr.save()
-        topic_2_en = taxonomies_factory.ResearchTopicFactory()
-        translation.activate(self.fr_locale.language_code)
-
-        response = self.client.get(self.library_page.localized.url)
-
-        topic_option_values = [i["value"] for i in response.context["topic_options"]]
-        self.assertEqual(len(topic_option_values), 2)
-        self.assertNotIn(topic_1_en.id, topic_option_values)
-        self.assertIn(topic_1_fr.id, topic_option_values)
-        self.assertIn(topic_2_en.id, topic_option_values)
-
     def test_filter_topic(self):
         topic_A = taxonomies_factory.ResearchTopicFactory()
         detail_page_A = detail_page_factory.ResearchDetailPageFactory(
@@ -668,57 +545,6 @@ class TestResearchLibraryPage(research_test_base.ResearchHubTestCase):
         self.assertNotIn(detail_page_1, research_detail_pages)
         self.assertNotIn(detail_page_2, research_detail_pages)
 
-    def test_research_regions_in_options(self):
-        region_1 = taxonomies_factory.ResearchRegionFactory()
-        region_2 = taxonomies_factory.ResearchRegionFactory()
-
-        response = self.client.get(self.library_page.url)
-
-        region_option_values = [i["value"] for i in response.context["region_options"]]
-        self.assertEqual(len(region_option_values), 2)
-        self.assertIn(region_1.id, region_option_values)
-        self.assertIn(region_2.id, region_option_values)
-
-    def test_region_in_options_matches_active_locale(self):
-        region_en = taxonomies_factory.ResearchRegionFactory()
-        region_fr = region_en.copy_for_translation(self.fr_locale)
-        region_fr.save()
-
-        response_en = self.client.get(self.library_page.localized.url)
-        translation.activate(self.fr_locale.language_code)
-        response_fr = self.client.get(self.library_page.localized.url)
-
-        region_option_values_en = [i["value"] for i in response_en.context["region_options"]]
-        self.assertEqual(len(region_option_values_en), 1)
-        self.assertIn(region_en.id, region_option_values_en)
-        self.assertNotIn(region_fr.id, region_option_values_en)
-        region_option_values_fr = [i["value"] for i in response_fr.context["region_options"]]
-        self.assertEqual(len(region_option_values_fr), 1)
-        self.assertNotIn(region_en.id, region_option_values_fr)
-        self.assertIn(region_fr.id, region_option_values_fr)
-
-    def test_localized_region_options(self):
-        """
-        Use active locales version of region if available.
-
-        If no translation is available for a given region, display the default locale
-        region.
-
-        """
-        region_1_en = taxonomies_factory.ResearchRegionFactory()
-        region_1_fr = region_1_en.copy_for_translation(self.fr_locale)
-        region_1_fr.save()
-        region_2_en = taxonomies_factory.ResearchRegionFactory()
-        translation.activate(self.fr_locale.language_code)
-
-        response = self.client.get(self.library_page.localized.url)
-
-        region_option_values = [i["value"] for i in response.context["region_options"]]
-        self.assertEqual(len(region_option_values), 2)
-        self.assertNotIn(region_1_en.id, region_option_values)
-        self.assertIn(region_1_fr.id, region_option_values)
-        self.assertIn(region_2_en.id, region_option_values)
-
     def test_filter_region(self):
         region_A = taxonomies_factory.ResearchRegionFactory()
         detail_page_A = detail_page_factory.ResearchDetailPageFactory(
@@ -798,26 +624,6 @@ class TestResearchLibraryPage(research_test_base.ResearchHubTestCase):
         self.assertIn(detail_page_2_fr, research_detail_pages)
         self.assertNotIn(detail_page_1, research_detail_pages)
         self.assertNotIn(detail_page_2, research_detail_pages)
-
-    def test_years_in_options(self):
-        year_1 = timezone.now().year
-        year_2 = year_1 - 1
-        detail_page_factory.ResearchDetailPageFactory(
-            parent=self.library_page,
-            original_publication_date=datetime.date(year=year_1, month=1, day=1),
-        )
-        detail_page_factory.ResearchDetailPageFactory(
-            parent=self.library_page,
-            original_publication_date=datetime.date(year=year_2, month=1, day=1),
-        )
-
-        response = self.client.get(self.library_page.url)
-
-        year_option_values = [i["value"] for i in response.context["year_options"]]
-        # It's 3 options because of the two years and the "any" option.
-        self.assertEqual(len(year_option_values), 3)
-        self.assertIn(year_1, year_option_values)
-        self.assertIn(year_2, year_option_values)
 
     def test_filter_for_year(self):
         detail_page_1 = detail_page_factory.ResearchDetailPageFactory(
